Rename FO4Archive to BA2Archive, merge SFArchive and FO4Archive
