+++ Wabbajack.Compression.BSA/BA2Archive/IBA2FileEntry.cs
@@ -1,6 +1,6 @@
 using Wabbajack.Compression.BSA.Interfaces;
 
-namespace Wabbajack.Compression.BSA.FO4Archive;
+namespace Wabbajack.Compression.BSA.BA2Archive;
 
 internal interface IBA2FileEntry : IFile
 {
