+++ Wabbajack.Compression.BSA/BA2Archive/IFileBuilder.cs
@@ -2,7 +2,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Wabbajack.Compression.BSA.FO4Archive;
+namespace Wabbajack.Compression.BSA.BA2Archive;
 
 internal interface IFileBuilder
 {
