+++ Wabbajack.Compression.BSA/BA2Archive/TextureChunk.cs
@@ -1,6 +1,6 @@
 using System.IO;
 
-namespace Wabbajack.Compression.BSA.FO4Archive;
+namespace Wabbajack.Compression.BSA.BA2Archive;
 
 public class TextureChunk
 {
